Fixed sql syntax

+++ dags/import_edinburgh_rightmove.py
@@ -25,7 +25,7 @@ with dag:
 DROP TABLE IF EXISTS landing.edinburgh{{ ds_nodash }}; 
 CREATE TABLE landing.edinburgh{{ ds_nodash }}
 (
-ID int not null, 
+ID bigint not null, 
 address varchar(128) not null, 
 number_of_beds smallint null, 
 links varchar(256) not null,
